REFACTOR: Move the arcade module to a separate crate to aid in conditional compilation and doing more platform-specific behavior
