+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 members = [
     "crates/thetawave_interface",
     "crates/thetawave_storage",
+    "crates/thetawave_arcade",
 ]
 
 [workspace.dependencies]
@@ -38,14 +39,15 @@ bevy_kira_audio = { version = "0.16.0", features = ["mp3", "wav"] }
 #bevy_editor_pls = "0.4.0"
 winit = "0.28.4"
 image = "0.24.6"
-bytes = "1.4.0"
+
 thetawave_interface = { path = "crates/thetawave_interface" }
 
 
 # Serial ports do not work in in the browser.
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-bevy_serialport = {version = "0.4.0", optional = true}
 thetawave_storage = { path = "crates/thetawave_storage", optional = true }
+thetawave_arcade = { path = "crates/thetawave_arcade", optional = true }
+
 
 # optimize dev packages as we don't need them in debug version
 [profile.dev.package."*"]
@@ -59,5 +61,5 @@ opt-level = 1
 codegen-units = 1
 
 [features]
-arcade = ["bevy_serialport"]
+arcade = ["thetawave_arcade"]
 storage = ["thetawave_storage"]
