+++ crates/thetawave_arcade/src/lib.rs
@@ -1,0 +1,2 @@
+/// Exposes a single Bevy plugin for integrating the arcade IO into the game systems.
+pub mod plugin;
